update info modal to accomodate any kind of html not just strings
